new test for animation -- not complete yet.
