+++ py_gd/test/test_gd.py
@@ -1268,227 +1268,3 @@ def test_clip_draw():
     img.save(outfile(fname))
 
     assert check_file(fname)
-
-
-def rotating_line(size=200):
-    """
-    generate the endpoints of a rotating line
-
-    for use in animation tests
-
-    it's expecting to be used in a square image:
-
-    (size,size)
-    """
-    endpoints = np.array(((-size / 2, 0), (size / 2, 0)))
-    offset = np.array((size / 2, size / 2))
-
-    for ang in range(0, 360, 10):
-        rad = np.deg2rad(ang)
-        rot_matrix = [(np.cos(rad), np.sin(rad)), (-np.sin(rad), np.cos(rad))]
-        points = np.dot(endpoints, rot_matrix).astype(np.int32) + offset
-        yield points
-
-
-def test_animation():
-    img = Image(200, 200)
-
-    anim = Animation(outfile("test_animation.gif"))
-
-    anim.begin_anim(img, 0)
-
-    for points in rotating_line(200):
-        img.draw_line(points[0], points[1], 'red')
-        anim.add_frame(img)
-
-    img.draw_line(np.array((0, 100)), np.array((200, 100)), 'green')
-    anim.add_frame(img)
-
-    anim.close_anim()
-
-    # not much to auto-check here
-    print(f"{anim.frames_written} frames were written")
-    assert anim.frames_written == 22
-
-
-def test_static_animation():
-    """
-    If subsequent frames are identical, then it should add tot he delay,
-    rather than adding duplicate images
-
-    Not sure how to actually test the delay, but looking at the animation
-    you can tell it's slower than the previous test one, which is otherwise
-    the same
-    """
-    img1 = Image(200, 200)
-    img2 = Image(200, 200)
-
-    anim = Animation(outfile("test_animation_static.gif"))
-    anim.begin_anim(img1, 0)
-
-    for points in rotating_line(200):
-        img1.draw_line(points[0], points[1], 'red')
-        img2.draw_line(points[0], points[1], 'red')
-
-        assert img1 == img2
-
-        anim.add_frame(img1)
-        anim.add_frame(img2)
-
-    anim.close_anim()
-    print(f"{anim.frames_written} frames were written")
-    # duplicate images should have added to delay, rather than adding an image
-    assert anim.frames_written == 21
-
-
-def test_animation_reuse_filename():
-    """
-    make an animation, then make another one with the same filename
-
-    The final one should be green lines
-
-    NOTE: we were having issues on Windows with permissions
-    """
-
-    for color in ('red', 'green'):
-        img = Image(200, 200)
-
-        anim = Animation(outfile("test_animation_reuse.gif"))
-        anim.begin_anim(img, 0)
-
-        for points in rotating_line(200):
-            img.draw_line(points[0], points[1], color, line_width=3)
-            anim.add_frame(img)
-        anim.close_anim()
-
-    # not much to auto-check here
-    print(f"{anim.frames_written} frames were written")
-    assert anim.frames_written == 21
-
-
-def test_animation_reuse_filename_not_close():
-    """
-    make an animation, then make another one with the same filename
-
-    The final one should be blue lines
-
-    NOTE: we were having issues on Windows with permissions
-    """
-
-    for color in ('red', 'blue'):
-        img = Image(200, 200)
-
-        anim = Animation(outfile("test_animation_reuse_not_close.gif"))
-        anim.begin_anim(img, 0)
-
-        for points in rotating_line(200):
-            img.draw_line(points[0], points[1], color, line_width=3)
-            anim.add_frame(img)
-        # anim.close_anim()
-
-    # not much to auto-check here
-    print(f"{anim.frames_written} frames were written")
-    assert anim.frames_written == 21
-
-
-def test_animation_reset_new_filename():
-    """
-    create an animation, then reset and make another one
-    """
-    anim = Animation(outfile("test_animation_reset1.gif"))
-    img = Image(200, 200)
-
-    anim.begin_anim(img, 0)
-
-    for points in rotating_line(200):
-        img.draw_line(points[0], points[1], 'red', line_width=3)
-        anim.add_frame(img)
-
-    assert anim.frames_written == 21
-
-    anim.reset(file_path=outfile("test_animation_reset2.gif"))
-    assert anim.frames_written == 0
-
-    img = Image(300, 300)
-    anim.begin_anim(img, 0)
-
-    for points in rotating_line(300):
-        img.draw_line(points[0], points[1], 'blue', line_width=3)
-        anim.add_frame(img)
-
-    anim.close_anim()
-
-    # not much to auto-check here
-    print(f"{anim.frames_written} frames were written")
-    assert anim.frames_written == 21
-
-
-def test_animation_reset_same_filename():
-    """
-    create an animation, then reset and make another one
-    using the same filename (by default)
-    """
-    anim = Animation(outfile("test_animation_reset_same.gif"))
-    img = Image(200, 200)
-
-    anim.begin_anim(img, 0)
-
-    for points in rotating_line(200):
-        img.draw_line(points[0], points[1], 'red', line_width=3)
-        anim.add_frame(img)
-
-    assert anim.frames_written == 21
-
-    anim.reset()
-    assert anim.frames_written == 0
-
-    img = Image(300, 300)
-    anim.begin_anim(img, 0)
-
-    for points in rotating_line(300):
-        img.draw_line(points[0], points[1], 'blue', line_width=3)
-        anim.add_frame(img)
-
-    anim.close_anim()
-
-    # not much to auto-check here
-    print(f"{anim.frames_written} frames were written")
-    assert anim.frames_written == 21
-
-
-def test_animation_delete_before_use():
-    """
-    make sure the dealloc doesn't barf if the file hasn't
-    been opened yet
-    """
-
-    filename = outfile("nothing.gif")
-
-    filename.unlink(missing_ok=True)
-    assert not filename.exists()
-
-    anim = Animation(filename)
-    del anim
-    assert not filename.exists()
-
-    # note: this creates a broken gif
-    anim = Animation(filename)
-    img = Image(200, 200)
-    anim.begin_anim(img, 0)
-    del anim
-    assert filename.exists()
-
-
-def test_animation_delete_one_frame():
-    """
-    make sure the dealloc creates a valid gif with only one frame added
-    """
-    filename = outfile("one_frame_delete.gif")
-    anim = Animation(filename)
-    img = Image(200, 200)
-    anim.begin_anim(img, 0)
-    img.draw_line((0, 0), (200, 200), color="white", line_width=4)
-    anim.add_frame(img)
-    del anim
-
-    assert filename.exists()
